Move chdir context manager into the srpm module
This is the only consumer of this method, so having it in the same file
is easier.
